fix(sound): 任务栏音量插件与控制中心音量模块音量数值不匹配
double类型转int精度丢失导致

Log: 任务栏音量插件与控制中心音量模块音量数值不匹配
https: //pms.uniontech.com/zentao/bug-view-16696.html
Reviewed-on: http://gerrit.uniontech.com/c/dde-control-center/+/469
Tested-by: yekaisheng <yekaisheng@uniontech.com>

+++ src/frame/window/modules/sound/speakerpage.cpp
@@ -130,9 +130,9 @@ void SpeakerPage::initSlider()
     //当底层数据改变后，更新滑动条显示的数据
     connect(m_model, &SoundModel::speakerVolumeChanged, this, [ = ](double v) {
         m_speakSlider->blockSignals(true);
-        m_speakSlider->setValue(static_cast<int>(v * 100));
+        m_speakSlider->setValue(v * 100);
         m_speakSlider->blockSignals(false);
-        m_outputSlider->setValueLiteral(QString::number(int(v * 100)) + "%");
+        m_outputSlider->setValueLiteral(QString::number(v * 100) + "%");
     });
 
     connect(m_model, &SoundModel::maxUIVolumeChanged, this, [ = ](double maxvalue) {
